requirements.txt
PyGitHub==1.51
termcolor==1.1.0
dulwich==0.20.5
jinja2==2.11.2